# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/updateManufacturer/{id}": {
      "post": {
        "tags": [
          "manufacturer-controller"
        ],
        "operationId": "updateManufacturer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManufacturerVO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/updateBeer/{id}": {
      "post": {
        "tags": [
          "beer-controller"
        ],
        "operationId": "updateBeer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerVO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/addManufacturer": {
      "post": {
        "tags": [
          "manufacturer-controller"
        ],
        "operationId": "addManufacturer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManufacturerVO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/addBeer": {
      "post": {
        "tags": [
          "beer-controller"
        ],
        "operationId": "addBeer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeerVO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/manufacturers": {
      "get": {
        "tags": [
          "manufacturer-controller"
        ],
        "operationId": "searchManufacturer",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Manufacturer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/manufacturer/{id}": {
      "get": {
        "tags": [
          "manufacturer-controller"
        ],
        "operationId": "getManufacturerDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Manufacturer"
                }
              }
            }
          }
        }
      }
    },
    "/beers": {
      "get": {
        "tags": [
          "beer-controller"
        ],
        "operationId": "getAllBeers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeerVO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/beer/{id}": {
      "get": {
        "tags": [
          "beer-controller"
        ],
        "operationId": "getBeerDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Beer"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ManufacturerVO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          }
        }
      },
      "BeerVO": {
        "type": "object",
        "properties": {
          "manufacturerName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "graduation": {
            "type": "string"
          }
        }
      },
      "Beer": {
        "type": "object",
        "properties": {
          "beerId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "graduation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/Manufacturer"
          }
        }
      },
      "Manufacturer": {
        "type": "object",
        "properties": {
          "manufacturerId": {
            "type": "integer",
            "format": "int64"
          },
          "beers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Beer"
            }
          },
          "company": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          }
        }
      }
    }
  }
}